ensure correct type is used in 64bit code

+++ ThrEd4/form.cpp
@@ -1301,7 +1301,7 @@ void form::chkseq(bool border) {
 void form::internal::ritbrd(unsigned& interleaveSequenceIndex2) {
 	if (SequenceIndex != 0u) {
 		InterleaveSequenceIndices[interleaveSequenceIndex2]
-		    = INSREC{ TYPBRD, gsl::narrow<unsigned int>(SelectedForm->borderColor) & COLMSK, InterleaveSequence->size(), I_BRD };
+		    = INSREC{ TYPBRD, gsl::narrow<unsigned int>(SelectedForm->borderColor) & COLMSK, gsl::narrow<unsigned int>(InterleaveSequence->size()), I_BRD };
 		form::chkseq(true);
 		interleaveSequenceIndex2++;
 	}
@@ -1310,7 +1310,7 @@ void form::internal::ritbrd(unsigned& interleaveSequenceIndex2) {
 void form::internal::ritapbrd(unsigned& interleaveSequenceIndex2) {
 	if (SequenceIndex != 0u) {
 		InterleaveSequenceIndices[interleaveSequenceIndex2]
-		    = INSREC{ TYPMSK, gsl::narrow<unsigned int>(SelectedForm->borderColor) >> 4, InterleaveSequence->size(), I_AP };
+		    = INSREC{ TYPMSK, gsl::narrow<unsigned int>(SelectedForm->borderColor) >> 4, gsl::narrow<unsigned int>(InterleaveSequence->size()), I_AP };
 		form::chkseq(true);
 		interleaveSequenceIndex2++;
 	}
@@ -1319,7 +1319,7 @@ void form::internal::ritapbrd(unsigned& interleaveSequenceIndex2) {
 void form::internal::ritfil(unsigned& interleaveSequenceIndex2) {
 	if (SequenceIndex != 0u) {
 		InterleaveSequenceIndices[interleaveSequenceIndex2]
-		    = INSREC{ TYPFRM, SelectedForm->fillColor, InterleaveSequence->size(), I_FIL };
+		    = INSREC{ TYPFRM, SelectedForm->fillColor, gsl::narrow<unsigned int>(InterleaveSequence->size()), I_FIL };
 		form::chkseq(false);
 		interleaveSequenceIndex2++;
 	}
